Fix low priority issues from code review

+++ src/Algorithms.Specs/StringDistance/StringSift3_Tests/When_comparing_two_very_similar_words.cs
@@ -8,7 +8,7 @@
     public class When_comparing_two_very_similar_words
     {
         private static StringSift3 sift;
-        private static float distance;
+        private static double distance;
 
         private Establish context = () => { sift = new StringSift3(); };
